[connman] Added iptables save and clear functionality that can be called via D-Bus. Fixed iptables protocol and match handling.
Most of the code is taken from iptables.c and iptables-save.c in iptables src.

Saves ipt in same format as iptables-save command to /etc/iptables/rules.v4. If
directory /etc/iptables does not exist, it will be created. Permissions are set
to 700 for directory and 600 for file. Checks if the given path is leading elsewhere
or is a symlink or if the file at given path is executable. Currently allows to
write only to /etc/iptables.

Clear should be done only for the "filter" (or own specific table to avoid accessing
connman specific tables and chains. TODO: implement own sailfish-tables/chains and
prevent clearing of all (?).

Added support to connman iptables.c for protocol and proper handling of matches
with protocol defined.

TODO: implement loading of iptables from a file.

Cleaned up the exposed api.