[connman] Added feature into sailfish iptables extension to support retrieval of iptables content.
Iptables content is saved to two GLists: chains and rules. Chains are saved as "CHAIN POLICY" and rules are saved as is (no parsing). This created structure should be free'd with connman_iptables_clear_content().

Small fix also to chain policy setting; if chain is not builtin the policy is skipped as iptc_set_policy supports only builtin chains.