ntp: Add callback to __connman_ntp_start()
Introduce a callback to notify timeserver if an operation was
successfull or not. The NTP layer should not be involved in the logic
of timerserver selection.

While at it, we also notify timerserver of any failure during decoding
the received NTP frame. Until now we just did that when we were not
able to communicate with the peer. The timerserver code is now able to
select the next server.

(cherry picked from commit fd54478617fe26b0e24affaf9ec235c413beef14)