[connman] Sailfish iptables API support. Contributes to JB#39338
Adds support for managing iptables over D-BUS with the help of a Connman
external plugin. Iptables API support is implemented as internal plugin
that can be enabled with --enable-sailfish-iptables-extension and is
located in sailfish_iptables_extension.c. The header to include
(iptables_extension.h) is installed as part of devel package.

Functionalities implemented for plugins to use:
 - Chain management (new, delete, flush, find)
 - Iptables rule management (insert, append, delete)
 - Commit iptables changes
 - Change iptables policy
 - Save, load and clear iptables
    - default file STORAREDIR/iptables/rules.v4 to store rules
    - Saving of iptables feature is adapted from iptables source.
 - Get iptables content

These functionalities do not restrict use of specific iptables table.
Saving and loading should be requested by the external plugin.

Upgrade iptables build requirement and install requirement to at least 1.6.1.