[connman-vpn] Connect VPN after vpnd crash if autoconnect set. JB#41684
This commit adds function for checking if autoconnect is set on a provider.

This fixes the issue that VPN having autoconnect on is not connected
after vpnd has crashed and started again. Added check for autoconnect in
plugins/vpn.c the in same place as pending connection has been checked,
where added provider will be connected.