[connman-vpn] Prevent connecting VPN if not online. Fixes JB#41705
This commit fixes the following:
 - Do not allow VPNs to be connected if there is no network
 - Do not leave VPNs connected when networks are disconnected one after
   another very quickly leaving VPN connection initialized due to
   default service change
 - Force VPNs to go through the default service always

This adds ConnMan state listener to vpnd. It first calls GetProperties
method of net.connman.Manager interface to get the state. If state is
not received vpnd does not accept VPNs to be connected until a signal
PropertyChanged with the ConnMan State has been received and the value
of the State is either "ready" or "online".

This adds ConnMan service listener to vpnd. When ConnMan is disconnected
the state is set to offline. When ConnMan is connected or vpnd is
started a function is added to glib main event loop with 500ms interval.

Connman state is tracked within vpnd with a boolean. Value "true" =
online/ready, value "false" = offline/idle.

This also adds a feature to disconnect all VPNs if the default changes
to NULL (no network) or when the default service changes in order to all
VPNs to use the default connection at all times. The disconnection and
autoconnection of VPNs happens only if the service type is not VPN.