[connman-vpn] Check if agent is present at VPN connect. Fixes JB#41892
This commit fixes the issue of not being able to autoconnect VPNs that
require credentials for authentication. This was caused by a timing
issue: (VPN) agents were not registered yet through VPN manager D-Bus
interface when autoconnect for a VPN was called at boot. Now the
connection of VPNs is delayed until a default agent is being set.

This commit also includes cleanup for vpn-provider.c since initial state
of connman is not needed to be explicitely requested in
__vpn_provider_init() because g_dbus_add_service_watch() calls the
connect() callback function if ConnMan service is present in D-Bus.
connman_service_watch_connected() is called then and requests the state
when the listeners are reqistered.

Also the indication when the state request has been done is moved to
after the state reply is processed to indicate that state is requested
and reply is processed thus, meaning that the whole process is completed
(hence the name change of the variable).

Added use of G_SOURCE_CONTINUE and G_SOURCE_REMOVE macros as return
values to timeout functions for readability.