Bookkeeping of VPN connections. Contributes to JB#41939
Added bookkeeping of VPN connections to track which transport service
was used to connect a VPN service. The pointer to the current default
service recorded for each connecting VPN and if a service is being
disconnected all the VPNs using the technique with that index are
disconnected.

Improved default_changed() behavior to disconnect the depending VPN
services when default is changed, or to disconnect the VPN service if it
was the previous default. If the new default service is depending on the
previous default service VPNs are not disconnected to prevent
connect-disconnect loops.