[connman] Improve default service selection logic. Contributes to JB#41939
This commit improves the logic to select the default service. The order
of the technologies is used from the preferred techs list. If the
service is higher on the list (index is smaller) and the new service is
lower in the list (index is bigger) the new service will not be set as
default.

With VPN being the default the dependency of the VPN is checked that if
the new service outranks the current transport service of the VPN being
as default service. In such case where new service, e.g. WiFi is
connected and VPN uses cellular, VPN will be disconnected and
reconnected over WiFi.

Added also more vpn autoconnection in case the default service is not
changed. Also if the default service is being set NULL and there is some
service that is connected that connected service is used instead of
NULL. Furthermore, resolver is called to redo the dns servers list if
the service list is modified in default_changed().