[connman] Use DNS of depending non-default route VPN. Contributes to JB#42337
Added a functionality to the dnsproxy which includes the DNS servers of
the VPN that depends on the current default service. The DNS servers of
a VPN are included only when it is not used as default route.

Set to do default change notification (in src/service.c:default_changed())
even if the default service was not changed. This is because DNS servers
of the VPN that is not being set as default route may not have been set
or enabled.