[connman] Improve service autoconnection. Contributes to JB#42337
When setting a NULL default service call autoconnection of services.
This is important since when dropping out of range of a WLAN network,
for instance, the mobile data should be brought up as quick as possible
if available.

Also added more checks into auto_connect_service() that when the active
count is 0 it does not always mean that autoconnection of services can
be stopped. The scenario where WLAN network is lost and mobile data is
on but exists as a idle service in the services list the mere check for
active_count amount is not enough since there may be a preferred tech
in the list that should be connected. Thus, a check for preferred
requirement and if such tech has been found is added. The above scenario
would result in some cases a delay of ~90s when dropping out from WLAN
and getting the mobile data online again.