[connman] Increase service online state checks. Contributes to JB#42337
Added a function to check if service is in online state.

Improved checking of default service selection in
allow_service_over_default() by allowing a service to be set over a
non-connected default service. Also, if the transport service of a VPN
that is current default is online but the new service is not online the
change is not made. Additionally, if the current default is online but
new service is not the change is not allowed.