[connman-vpn] Improve auto connection of VPNs. Contributes to JB#42337
Added more checks for run_vpn_auto_connect(): 1) delay connecting of VPN
if the current default service is not yet online but in ready state and
2) stop VPN auto connection if there is no default service in online
state or the default service is connected VPN. This will be added to
main loop when there is a change in the default service.

Additionally keep the run_vpn_auto_connect() in main loop with 1s delay
after the first run in order to re-connect any VPN that should be
connected. Because of errors in network conditions the VPN can get
disconnected for a multiple reasons. By keeping the auto connect
function in the main loop the VPN that should be connected will be
attempted until user decides to disable it. If there is no VPN to
automatically connect increase run_vpn_auto_connect() delay to 2s. It is
kept in the main loop because if autoconnect is removed from a VPN that
is failing then the autoconnect loop would be never re-triggered.

Removed unsetting of VPN autoconnect in vpn/plugins/vpn.c:vpn_notify()
to make the toggling of autoconnect setting completely a user setting.
ConnMan should not disable it for the first error it encounters.

Added vpn_auto_connect() to service retrying. Now if the failed service
is a VPN the vpn_auto_connect() is called, otherwise
__connman_service_auto_connect() is called.