[connman] Route VPN DNS requests over VPN peer. Fixes JB#42766
Routes to DNS servers added by a VPN connection should use the peer
address of the VPN. Otherwise the route that is added to DNS of a VPN is
going over the transport service.

This fixes the issue by checking the service type for VPN and the type
of the peer address before the routes to DNS servers are added. If the
peer address of VPN is suitable it should be used for the VPN DNS
routes.