[connman-vpn] Set order for VPN created from provider. Fixes JB#42766
This fix adds the use of set_split_routing() when creating a new VPN
service from provider. Value for do_split_routing is initialized as
false when creating a new service struct. set_split_routing() sets the
service order for VPNs based on this value. When the VPN service is
created from provider it has no configuration file so
__connman_service_load_modifiable() cannot load any values.

This causes the issue of newly added VPN not being sorted on top of the
service list and it is not treated as default service. As a result of
this the DNS servers of the VPN are not enabled and all DNS traffic goes
over transport service DNS.

The reason why the __connman_service_load_modifiable() has worked before
in loading the values is because the VPN services are not actually
removed from service.c service_list when they are connected and removed
via UI. So in case a new VPN is added that user presumably has removed
via UI when the VPN was connected that had the same name then this
problem would have gone unnoticed.