[connman-vpn] Add remove functionality to VPN driver. Fixes JB#43005
This commit adds remove functionality to the VPN driver structure. The
prototype is int (remove*) (struct vpn_provider *provider) and is to be
registered within the VPN plugin.

This function is called from vpn/plugins/vpn.c when removing a provider.
It is meant for VPNs that need to cleanup some VPN specific content when
removed or to notify other services after removal.