vpn: Remove service data when VPN connection is removed.
This commit adds a call to connman_service_unref() when a signal from
connman-vpnd is received indicating that a VPN connection is removed.
This way all references to VPN connection removed via connman-vpn are
removed also from connmand.

Otherwise the VPN connection service data is kept in service.c
service_list and when adding the same VPN using, e.g., a VPN
configuration file the settings that were not overridden by the config
will be used from the service settings. This happens because each news
service addition includes a search for existing service in service_list.