[connman] Rework service removal process. Contributes to JB#42797
This commit changes the service removal process to comply a bit better
to the process in upstream. Removal process is now going through
connman_service_unref() which 1) reduces the reference count and if it
is zero it 2) removes the service from service_list, 3) disconnects the
service, 4) removes reference from service_hash which in turn 5) calls
service_free() to clear the service content that 6) calls the
service_schedule_removed() to indicate that this service is gone.

First of all, changing the service_hash hash table to use service_free()
as removal function. This is because the reference from service_hash table
is removed if unref function is used and reference count is 0. This is
the use in upstream (commit 02d474014e806475fbc56b0e9fbebc447d524fb9).

Second, the service_schedule_removed() can be called when
service_free() is used. Old implementation used to do this with another
function that complicates things and goes into different way as
upstream.

Third, when removing a service call connman_service_unref() instead of
service_remove(). It is better to also notify about these removals as
the service_schedule_removed() is eventually called if the reference
count is zero. Also, removal from service_list and service_hash is
handled by the unref function.

Fourth, functions service_removed() and service_remove() are not in
upstream and became useless with the changes above, are removed.