[connman] Add tests for managed iptables rules. JB#42675
Tests for the managed iptables rules read from firewall.conf are added
in this commit. The tests include installing a new set of rules using a
temporary firewall.conf (set in /tmp/connman_test), checking that rules
exist, checking that rules are added and removed when services go up and
down and that iptables is cleaned up properly after shutting down
(cleanup calls to iptables.c and firewall.c).

Also some invalid rules are checked. This list could be expanded later.

The temporary firewall.conf and the directory will be after tests
finish.