[connman] Implement firewall failsafe mechanism. Contributes to JB#43998
This commit adds a failsafe mechanism to set policies of all chains to
ACCEPT in case the setup of dynamic rules fails due to iptables error.
It will cleanup and initialize iptables.c before setting the policies.

Changed firewall rule enabling to report the last error and not to stop
when one error is encountered while enabling all rules.