[connman] Improve firewall iptables rule option parsing. JB#43924
This commit implements a better parser for all the supported iptables
rule (match) options. Without proper checks iptables will call the
exit_err() function that will exit the program if the option (in some
cases) or the value (in some cases) is invalid. For this reason a
parsing of invalid content is imperative.

Each match option has to also have proper protocol and match set.
Protocols are checked also using their int representation.

Options for iptables matches that are processed:
 - tcp
 - mark
 - conntrack (address switches not supported)
 - ttl
 - pkttype
 - limit
 - helper (no parsing as iptables does not care about value)
 - enc
 - ah
 - esp
 - mh (match -m mh does not work in connman, ignored)
 - sctp (match -m sctp does not work in connman, ignored)
 - icmp
 - icmpv6 && ipv6-icmp
 - dccp

Also added tests for iptables rule options. All accepted options are
checked as well as invalid options are attempted to be used with the
tests.