[connman] Improve firewall option checks. JB#43924
This commit introduces a hash table based approach in checking whether
the given option in iptables rule can be used with the specified match
or protocol. This reduces the search time as every possible option does
not need to be compared but only those that are supported by the match
or the protocol (with sctp and dccp match is not working).

Some, such as TCP, DCCP and multiport require checking from two separate
categories as the match (in case of DCCP the protocol) supports also key
"port" options in addition to its own options.

The hash table containing the iptables_type_option structs is
initialized at init and destroyed at cleanup.