[connman] Enable firewall for device using status notification. JB#43928
This commit adds function to enable and disable firewall configuration
for a device. Process is similar to tethering notification use. When
notified via notifier.c the interface from the struct connman_device is
used as hash table identifier as well as the interface for the rule.

The firewall rules contain allow all rule for the interface (incoming
traffic). Existing firewall for the device is reused if found, to reduce
memory fragmentation.

If the device is managed it is ignored. Rules for the managed devices
come from firewall configuration.