service: Track g_timeout for wispr try
In case we are disconnecting or removing the device,
the service will keep a ref until the timeout is reached.

Better to cancel the timeout, so that we can drop the ref,
and remove the service immediatly.

(cherry picked from commit d318b18d284b1be2cb3f490ffd5e722058bc5ad1)

---

ConnMan 1.37 upgrade comment:
 - Our fork has more commits on top of the wispr changes, these are old
   and unnecessary and most of the commit content was removed.