vpn: Check for VPN service validity if no online transports exist.
This commit fixes the following scenario:

    1) When connected to WiFi A connect manually to WiFi B
    2) Let VPN start to connect using WiFi B
    3) Shutdown WiFi B before VPN has connected
    4) Wait for connman to detect connection failure with WiFi B
    5) Let to connect to WiFi A and establish VPN connection
    6) There are 2 vpn adapters with stale routes of the old adapter

The connected VPN must be disconnected before default service is changed
and notified in the above scenario. By checking in transport validation
if the service is valid for the current scenario when no online
transport services exist it can be done within vpn.c plugin. Otherwise
it will later result unnecessay disconnect calls to VPN that is being
reconnected.