config: Add support matching on device name for provisioning
While matching on the MAC address is working for plain interfaces, it
wont work for managing VLAN interfaces. The VLAN interfaces share the
same MAC address with the parent interface.

The argument that the MAC address is the only stable way to identify a
device is a bit weak because the MAC can be changed via udev or other
means.

Furthermore, with systemd's feature of stable interface name it makes
things a lot easier for embedded system with a pre-provisioning
rootfs. Such devices have different MAC address but all of them have
the same interface name.

(cherry picked from commit a719719e126407004840e8da3b8139a7cfc98890)