[connman] Add connected services in preferred list. Contributes to JB#43191
Add also connected services in preferred technologies list for the
service_preferred_over() used in service sorting to work. Sorting of
services can happen in situation where change of network conditions is
quite fast, WiFi is being set down and VPN is used as autoconnected. It
manifests itself in situation where previous WiFi connection is abruplty
terminated (AP down) and VPN is connected over mobile data and then
there is new WiFi network detected and it is in ready state when
entering service_compare().

The mobile data service is in online state and WiFi is in ready state so
both are connected in service_compare() checks. These are then compared
using service_preferred_over() but WiFi has apparently no network yet as
this compare is done early right after transferring to ready state. The
list that is returned by preferred_tech_list_get() does not include the
WiFi service in ready state. Because of this, connected WiFi service
will remain below the mobile data service as it is included in the list.
And when it has VPN connected, the VPN will remain connected over mobile
data even after WiFi is enabled.

This issue apparently is because of some timing when the WiFi AP is
being put down and how fast the detection is done. It was reproducable
after fresh boot but did not reproduce after the device has been on for
a while and network changes have been done.

Adding the connected services also to the list returned by
preferred_tech_list_get() seems to solve the issue. The services that
are in ready state should be included in compare. It should not harm
other use as this is used in auto_connect_service() which will ignore
connecting and connected services anyways.