[connman] Fix service preference check. Contributes to JB#43191
Ignore VPN services in service preference check. VPNs are special case
depending on other services as transport. They are never in preferred
technologies lists.

Return the proper comparison value always. If either of the services has
index of -1 it is valid for comparison. The reason is that how the
current list is updated and when, for example, WiFi is on, mobile data
connections are not included in the list. In such cases, mobile data
services should be sorted below active WiFi services, which is obeyed
with this change.

Otherwise there would be cases when in case of moving out of WiFi range
or shutting down the WiFi connman starts to use mobile data and when VPN
is being used, in some autoconnection cases when another WiFi is found
and connected, it remains below mobile data. Reason is that WiFi
services are in preferred tech list and mobile data connections are not.
Because of this, 0 would have been returned and sorting would be
invalid. That aforementioned issue is fixed with this change.

Also reduce log noise on  __connman_service_is_default_route().